fix(docs): simplify examples
