+++ examples/schnorr.rs
@@ -1,11 +1,11 @@
-//! Example: simple Schnorr proof of knowledge using the sigma-rs framework.
+//! Example: Schnorr proof of knowledge.
 //!
-//! This example demonstrates how to prove knowledge of a discrete logarithm in a group $\mathbb{G}$ of prime order $p$ where the discrete logarithm problem is hard,
-//! using the Sigma protocol abstraction provided by `sigma-rs`.
+//! This example demonstrates how to prove knowledge of a discrete logarithm using `sigma-rs`.
+//! A common use case is authentication: proving you know a secret key without revealing it.
 //!
-//! More precisely, the prover convinces a verifier that it knows a secret $x$ such that: $$P = x \cdot G$$
+//! The prover convinces a verifier that it knows a secret $x$ such that: $$P = x \cdot G$$
 //!
-//! where $G$ is a fixed generator of $\mathbb{G}$ and $P$ is an element of the group.
+//! where $G$ is a generator of a prime-order group $\mathbb{G}$ and $P$ is a public group element.
 //!
 //! ---
 //!
@@ -48,50 +48,43 @@ use sigma_rs::schnorr_protocol::SchnorrProtocol;
 pub fn discrete_logarithm<G: Group + GroupEncoding>(
     x: G::Scalar,
 ) -> (LinearRelation<G>, Vec<G::Scalar>) {
-    let mut morphismp: LinearRelation<G> = LinearRelation::new();
+    let mut morphism: LinearRelation<G> = LinearRelation::new();
 
-    // === Allocation phase ===
-    // To construct a relation, we first create the variables (scalar, group points)
-    // that defines the relation. As they are variables, they are not real "values" (eg. a specific point of $\mathbb{Z}_p$).
+    // First, we allocate symbolic variables for our relation.
+    // These represent the structure of our proof, not concrete values yet.
 
     // Allocate a variable for the scalar witness `x`
-    let var_x = morphismp.allocate_scalar();
+    let var_x = morphism.allocate_scalar();
 
     // Allocate a variable for the group element `G` (the generator)
-    let var_G = morphismp.allocate_element();
+    let var_G = morphism.allocate_element();
 
-    // === Constraint phase ===
     // Now that all variables are defined, we describe the relation we want to prove.
     // The variable `P` is allocated now, as it is the result of the computation to be proven.
 
     // Create the constraint `P = x * G`
-    let var_P = morphismp.allocate_eq(var_x * var_G);
+    let var_P = morphism.allocate_eq(var_x * var_G);
 
-    // === Assignation phase ===
     // We now assign real values to the variables we have defined for our relation above.
     // Since the witness is provided to the function, we only need to assign the group points.
 
     // Assign the group generator to the corresponding variable `G`
-    morphismp.assign_element(var_G, G::generator());
+    morphism.assign_element(var_G, G::generator());
 
     // Assign the value of the image to the variable `P` (i.e., evaluate the group equation for `x`)
-    morphismp.compute_image(&[x]).unwrap();
+    morphism.compute_image(&[x]).unwrap();
 
-    // === Output phase ===
     // The relation has been defined and is ready to be used in a protocol.
-
     // Sanity check: ensure `P = x * G`
-    let P = morphismp.morphism.group_elements.get(var_P).unwrap();
+    let P = morphism.morphism.group_elements.get(var_P).unwrap();
     assert_eq!(P, G::generator() * x);
 
     // Output the relation and the witness for the upcoming proof
-    (morphismp, vec![x])
+    (morphism, vec![x])
 }
 
 #[allow(non_snake_case)]
 fn main() {
-    // === Step 1: Setup the relation and witness ===
-
     // Choose a secret scalar `x` (the witness).
     let x = Scalar::random(&mut OsRng);
 
@@ -100,8 +93,6 @@ fn main() {
     // - The witness in vector format (useful for proofs with multiple witnesses).
     let (relation, witness) = discrete_logarithm(x);
 
-    // === Step 2: Create the Schnorr protocol and wrap it with Fiat-Shamir ===
-
     // Build the Sigma protocol instance from the relation.
     let schnorr = SchnorrProtocol::<RistrettoPoint>::from(relation);
 
@@ -109,14 +100,10 @@ fn main() {
     // A domain separator is given as a byte-sequence to identify the current instance being proven.
     let nizk = NISigmaProtocol::<_, ShakeCodec<RistrettoPoint>>::new(b"schnorr-example", schnorr);
 
-    // === Step 3: Generate the proof ===
-
     // Generate a non-interactive proof with the witness.
     // The non-interactive proof contains the 3 elements of a sigma protocol transcript: a commitment, a challenge, and a response.
     // This transcript can be transmitted by the prover to any verifier, without revealing any secret information about the prover.
     let (commitment, challenge, response) = nizk.prove(&witness, &mut OsRng).unwrap();
-
-    // === Step 4: Verify the proof ===
 
     // Verification requires checking the proof against the transcript and constraint system.
     // The verifier can be convinced that the prover indeed knows `x` so that `P = x*G`. This doesn't require any interaction with the prover other than receiving the transcript.
